Manual backport of PR 468
