+++ tests/jenkins/jobs/BenchmarkTestMultinode_Jenkinsfile.txt
@@ -5,13 +5,15 @@
             BenchmarkTestMultinode_Jenkinsfile.script(groovy.lang.Closure)
                BenchmarkTestMultinode_Jenkinsfile.downloadBuildManifest({url=test://artifact.url, path=tests/data/opensearch-1.3.0-bundle.yml})
                   downloadBuildManifest.legacySCM(groovy.lang.Closure)
-                  downloadBuildManifest.library({identifier=jenkins@6.7.2, retriever=null})
-                  downloadBuildManifest.sh(curl -sSL test://artifact.url --output tests/data/opensearch-1.3.0-bundle.yml)
+                  downloadBuildManifest.library({identifier=jenkins@main, retriever=null})
+                  downloadBuildManifest.sh(mkdir -p $(dirname "tests/data/opensearch-1.3.0-bundle.yml"))
+                  downloadBuildManifest.sh(touch tests/data/opensearch-1.3.0-bundle.yml)
+                  downloadBuildManifest.sh(curl -sSL --retry 5 test://artifact.url --output tests/data/opensearch-1.3.0-bundle.yml)
                   downloadBuildManifest.readYaml({file=tests/data/opensearch-1.3.0-bundle.yml})
                   BuildManifest.asBoolean()
                BenchmarkTestMultinode_Jenkinsfile.runBenchmarkTestScript({bundleManifest=tests/data/opensearch-1.3.0-bundle.yml, distributionUrl=, distributionVersion=, workload=nyc_taxis, insecure=false, singleNode=false, minDistribution=false, use50PercentHeap=true, enableRemoteStore=true, suffix=307-secure, managerNodeCount=3, dataNodeCount=3, clientNodeCount=, ingestNodeCount=, mlNodeCount=, dataInstanceType=, userTag=key1:value1,security-enabled:true, workloadParams={"key2":"value2"}, testProcedure=custom-test-procedure, excludeTasks=index-append,default, includeTasks=type:search,index, additionalConfig=cluster.indices.replication.strategy:SEGMENT, dataStorageSize=200, mlStorageSize=200, jvmSysProps=, captureNodeStat=true, captureSegmentReplicationStat=false, telemetryParams={"telemetry_setting":"value"}})
                   runBenchmarkTestScript.legacySCM(groovy.lang.Closure)
-                  runBenchmarkTestScript.library({identifier=jenkins@6.7.2, retriever=null})
+                  runBenchmarkTestScript.library({identifier=jenkins@main, retriever=null})
                   runBenchmarkTestScript.readYaml({file=tests/data/opensearch-1.3.0-bundle.yml})
                   BuildManifest.asBoolean()
                   runBenchmarkTestScript.string({credentialsId=jenkins-aws-account-public, variable=AWS_ACCOUNT_PUBLIC})
@@ -33,13 +35,15 @@
             BenchmarkTestMultinode_Jenkinsfile.script(groovy.lang.Closure)
                BenchmarkTestMultinode_Jenkinsfile.downloadBuildManifest({url=test://artifact.url, path=tests/data/opensearch-1.3.0-bundle.yml})
                   downloadBuildManifest.legacySCM(groovy.lang.Closure)
-                  downloadBuildManifest.library({identifier=jenkins@6.7.2, retriever=null})
-                  downloadBuildManifest.sh(curl -sSL test://artifact.url --output tests/data/opensearch-1.3.0-bundle.yml)
+                  downloadBuildManifest.library({identifier=jenkins@main, retriever=null})
+                  downloadBuildManifest.sh(mkdir -p $(dirname "tests/data/opensearch-1.3.0-bundle.yml"))
+                  downloadBuildManifest.sh(touch tests/data/opensearch-1.3.0-bundle.yml)
+                  downloadBuildManifest.sh(curl -sSL --retry 5 test://artifact.url --output tests/data/opensearch-1.3.0-bundle.yml)
                   downloadBuildManifest.readYaml({file=tests/data/opensearch-1.3.0-bundle.yml})
                   BuildManifest.asBoolean()
                BenchmarkTestMultinode_Jenkinsfile.runBenchmarkTestScript({bundleManifest=tests/data/opensearch-1.3.0-bundle.yml, distributionUrl=, distributionVersion=, workload=nyc_taxis, insecure=true, singleNode=false, minDistribution=false, use50PercentHeap=true, enableRemoteStore=true, suffix=307, managerNodeCount=3, dataNodeCount=3, clientNodeCount=, ingestNodeCount=, mlNodeCount=, dataInstanceType=, userTag=key1:value1,security-enabled:false, workloadParams={"key2":"value2"}, testProcedure=custom-test-procedure, excludeTasks=index-append,default, includeTasks=type:search,index, additionalConfig=cluster.indices.replication.strategy:SEGMENT, dataStorageSize=200, mlStorageSize=200, jvmSysProps=, captureNodeStat=true, captureSegmentReplicationStat=false, telemetryParams={"telemetry_setting":"value"}})
                   runBenchmarkTestScript.legacySCM(groovy.lang.Closure)
-                  runBenchmarkTestScript.library({identifier=jenkins@6.7.2, retriever=null})
+                  runBenchmarkTestScript.library({identifier=jenkins@main, retriever=null})
                   runBenchmarkTestScript.readYaml({file=tests/data/opensearch-1.3.0-bundle.yml})
                   BuildManifest.asBoolean()
                   runBenchmarkTestScript.string({credentialsId=jenkins-aws-account-public, variable=AWS_ACCOUNT_PUBLIC})
