+++ tests/jenkins/jobs/BenchmarkTest_Jenkinsfile.txt
@@ -5,13 +5,15 @@
             BenchmarkTest_Jenkinsfile.script(groovy.lang.Closure)
                BenchmarkTest_Jenkinsfile.downloadBuildManifest({url=test://artifact.url, path=tests/data/opensearch-1.3.0-bundle.yml})
                   downloadBuildManifest.legacySCM(groovy.lang.Closure)
-                  downloadBuildManifest.library({identifier=jenkins@6.7.2, retriever=null})
-                  downloadBuildManifest.sh(curl -sSL test://artifact.url --output tests/data/opensearch-1.3.0-bundle.yml)
+                  downloadBuildManifest.library({identifier=jenkins@main, retriever=null})
+                  downloadBuildManifest.sh(mkdir -p $(dirname "tests/data/opensearch-1.3.0-bundle.yml"))
+                  downloadBuildManifest.sh(touch tests/data/opensearch-1.3.0-bundle.yml)
+                  downloadBuildManifest.sh(curl -sSL --retry 5 test://artifact.url --output tests/data/opensearch-1.3.0-bundle.yml)
                   downloadBuildManifest.readYaml({file=tests/data/opensearch-1.3.0-bundle.yml})
                   BuildManifest.asBoolean()
                BenchmarkTest_Jenkinsfile.runBenchmarkTestScript({bundleManifest=tests/data/opensearch-1.3.0-bundle.yml, distributionUrl=, distributionVersion=, workload=nyc_taxis, insecure=false, singleNode=true, minDistribution=false, use50PercentHeap=true, enableRemoteStore=false, suffix=307-secure, managerNodeCount=, dataNodeCount=, clientNodeCount=, ingestNodeCount=, mlNodeCount=, dataInstanceType=r5.8xlarge, userTag=security-enabled:true, workloadParams={"key2":"value2"}, testProcedure=, excludeTasks=, includeTasks=, additionalConfig=cluster.indices.replication.strategy:SEGMENT, dataStorageSize=200, mlStorageSize=200, jvmSysProps=, captureNodeStat=false, captureSegmentReplicationStat=true, telemetryParams=})
                   runBenchmarkTestScript.legacySCM(groovy.lang.Closure)
-                  runBenchmarkTestScript.library({identifier=jenkins@6.7.2, retriever=null})
+                  runBenchmarkTestScript.library({identifier=jenkins@main, retriever=null})
                   runBenchmarkTestScript.readYaml({file=tests/data/opensearch-1.3.0-bundle.yml})
                   BuildManifest.asBoolean()
                   runBenchmarkTestScript.string({credentialsId=jenkins-aws-account-public, variable=AWS_ACCOUNT_PUBLIC})
@@ -33,13 +35,15 @@
             BenchmarkTest_Jenkinsfile.script(groovy.lang.Closure)
                BenchmarkTest_Jenkinsfile.downloadBuildManifest({url=test://artifact.url, path=tests/data/opensearch-1.3.0-bundle.yml})
                   downloadBuildManifest.legacySCM(groovy.lang.Closure)
-                  downloadBuildManifest.library({identifier=jenkins@6.7.2, retriever=null})
-                  downloadBuildManifest.sh(curl -sSL test://artifact.url --output tests/data/opensearch-1.3.0-bundle.yml)
+                  downloadBuildManifest.library({identifier=jenkins@main, retriever=null})
+                  downloadBuildManifest.sh(mkdir -p $(dirname "tests/data/opensearch-1.3.0-bundle.yml"))
+                  downloadBuildManifest.sh(touch tests/data/opensearch-1.3.0-bundle.yml)
+                  downloadBuildManifest.sh(curl -sSL --retry 5 test://artifact.url --output tests/data/opensearch-1.3.0-bundle.yml)
                   downloadBuildManifest.readYaml({file=tests/data/opensearch-1.3.0-bundle.yml})
                   BuildManifest.asBoolean()
                BenchmarkTest_Jenkinsfile.runBenchmarkTestScript({bundleManifest=tests/data/opensearch-1.3.0-bundle.yml, distributionUrl=, distributionVersion=, workload=nyc_taxis, insecure=true, singleNode=true, minDistribution=false, use50PercentHeap=true, enableRemoteStore=false, suffix=307, managerNodeCount=, dataNodeCount=, clientNodeCount=, ingestNodeCount=, mlNodeCount=, dataInstanceType=r5.8xlarge, userTag=security-enabled:false, workloadParams={"key2":"value2"}, testProcedure=, excludeTasks=, includeTasks=, additionalConfig=cluster.indices.replication.strategy:SEGMENT, dataStorageSize=200, mlStorageSize=200, jvmSysProps=, captureNodeStat=false, captureSegmentReplicationStat=true, telemetryParams=})
                   runBenchmarkTestScript.legacySCM(groovy.lang.Closure)
-                  runBenchmarkTestScript.library({identifier=jenkins@6.7.2, retriever=null})
+                  runBenchmarkTestScript.library({identifier=jenkins@main, retriever=null})
                   runBenchmarkTestScript.readYaml({file=tests/data/opensearch-1.3.0-bundle.yml})
                   BuildManifest.asBoolean()
                   runBenchmarkTestScript.string({credentialsId=jenkins-aws-account-public, variable=AWS_ACCOUNT_PUBLIC})
