+++ tests/jenkins/jobs/PerfTest_Jenkinsfile.txt
@@ -5,13 +5,15 @@
             PerfTest_Jenkinsfile.script(groovy.lang.Closure)
                PerfTest_Jenkinsfile.downloadBuildManifest({url=test://artifact.url, path=tests/data/opensearch-1.3.0-bundle.yml})
                   downloadBuildManifest.legacySCM(groovy.lang.Closure)
-                  downloadBuildManifest.library({identifier=jenkins@6.7.2, retriever=null})
-                  downloadBuildManifest.sh(curl -sSL test://artifact.url --output tests/data/opensearch-1.3.0-bundle.yml)
+                  downloadBuildManifest.library({identifier=jenkins@main, retriever=null})
+                  downloadBuildManifest.sh(mkdir -p $(dirname "tests/data/opensearch-1.3.0-bundle.yml"))
+                  downloadBuildManifest.sh(touch tests/data/opensearch-1.3.0-bundle.yml)
+                  downloadBuildManifest.sh(curl -sSL --retry 5 test://artifact.url --output tests/data/opensearch-1.3.0-bundle.yml)
                   downloadBuildManifest.readYaml({file=tests/data/opensearch-1.3.0-bundle.yml})
                   BuildManifest.asBoolean()
                PerfTest_Jenkinsfile.runPerfTestScript({bundleManifest=tests/data/opensearch-1.3.0-bundle.yml, buildId=1236, architecture=x64, insecure=false, workload=nyc_taxis, testIterations=1, warmupIterations=1, stackNameSuffix=307})
                   runPerfTestScript.legacySCM(groovy.lang.Closure)
-                  runPerfTestScript.library({identifier=jenkins@6.7.2, retriever=null})
+                  runPerfTestScript.library({identifier=jenkins@main, retriever=null})
                   runPerfTestScript.readYaml({file=tests/data/opensearch-1.3.0-bundle.yml})
                   BuildManifest.asBoolean()
                   runPerfTestScript.sh(
@@ -31,13 +33,15 @@
             PerfTest_Jenkinsfile.script(groovy.lang.Closure)
                PerfTest_Jenkinsfile.downloadBuildManifest({url=test://artifact.url, path=tests/data/opensearch-1.3.0-bundle.yml})
                   downloadBuildManifest.legacySCM(groovy.lang.Closure)
-                  downloadBuildManifest.library({identifier=jenkins@6.7.2, retriever=null})
-                  downloadBuildManifest.sh(curl -sSL test://artifact.url --output tests/data/opensearch-1.3.0-bundle.yml)
+                  downloadBuildManifest.library({identifier=jenkins@main, retriever=null})
+                  downloadBuildManifest.sh(mkdir -p $(dirname "tests/data/opensearch-1.3.0-bundle.yml"))
+                  downloadBuildManifest.sh(touch tests/data/opensearch-1.3.0-bundle.yml)
+                  downloadBuildManifest.sh(curl -sSL --retry 5 test://artifact.url --output tests/data/opensearch-1.3.0-bundle.yml)
                   downloadBuildManifest.readYaml({file=tests/data/opensearch-1.3.0-bundle.yml})
                   BuildManifest.asBoolean()
                PerfTest_Jenkinsfile.runPerfTestScript({bundleManifest=tests/data/opensearch-1.3.0-bundle.yml, buildId=1236, architecture=x64, insecure=true, workload=nyc_taxis, testIterations=1, warmupIterations=1, stackNameSuffix=307})
                   runPerfTestScript.legacySCM(groovy.lang.Closure)
-                  runPerfTestScript.library({identifier=jenkins@6.7.2, retriever=null})
+                  runPerfTestScript.library({identifier=jenkins@main, retriever=null})
                   runPerfTestScript.readYaml({file=tests/data/opensearch-1.3.0-bundle.yml})
                   BuildManifest.asBoolean()
                   runPerfTestScript.sh(
