+++ vars/downloadBuildManifest.groovy
@@ -7,9 +7,10 @@
  * compatible open source license.
  */
 def call(Map args = [:]) {
-    def lib = library(identifier: "jenkins@6.7.2", retriever: legacySCM(scm))
-
-    sh "curl -sSL ${args.url} --output ${args.path}"
+    def lib = library(identifier: "jenkins@main", retriever: legacySCM(scm))
+    sh "mkdir -p \$(dirname \"${args.path}\")"
+    sh "touch ${args.path}"
+    sh "curl -sSL --retry 5 ${args.url} --output ${args.path}"
     def buildManifestObj = lib.jenkins.BuildManifest.new(readYaml(file: args.path))
     return buildManifestObj
 }
